Fix CIS issues (PipelineFileCache.cpp)
#rb none


#ROBOMERGE-SOURCE: CL 16693509 via CL 16694165 via CL 16694365
#ROBOMERGE-BOT: (v835-16672529)

[CL 16694420 by arciel rekman in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/RHI/Private/PipelineFileCache.cpp
@@ -551,9 +551,9 @@ void FPipelineCacheFileFormatPSO::GraphicsDescriptor::AddStateToReadableString(T
 	OutBuilder << TEXT(" MSAA:");
 	OutBuilder << MSAASamples;
 	OutBuilder << TEXT(" DSfmt:");
-	OutBuilder << DepthStencilFormat;
+	OutBuilder << uint32(DepthStencilFormat);
 	OutBuilder << TEXT(" DSflags:");
-	OutBuilder << DepthStencilFlags;
+	OutBuilder << uint32(DepthStencilFlags);
 	OutBuilder << TEXT("\n");
 
 	OutBuilder << TEXT(" DL:");
@@ -582,7 +582,7 @@ void FPipelineCacheFileFormatPSO::GraphicsDescriptor::AddStateToReadableString(T
 			OutBuilder << TEXT(":fmt=");
 			OutBuilder << uint32(RenderTargetFormats[Index]);
 			OutBuilder << TEXT(" flg=");
-			OutBuilder << RenderTargetFlags[Index];
+			OutBuilder << uint32(RenderTargetFlags[Index]);
 		}
 		OutBuilder << TEXT("\n");
 	}
